remove leva dependency from mesh-gradient page
