+++ docs/src/app/mesh-gradient/page.tsx
@@ -1,14 +1,13 @@
 'use client';
 
+import { useState } from 'react';
 import { MeshGradient, meshGradientPresets } from '@paper-design/shaders-react';
-import { useControls, button, folder } from 'leva';
-import { setParamsSafe, useResetLevaParams } from '@/helpers/use-reset-leva-params';
-import { usePresetHighlight } from '@/helpers/use-preset-highlight';
 import Link from 'next/link';
 import { BackButton } from '@/components/back-button';
-import { cleanUpLevaParams } from '@/helpers/clean-up-leva-params';
 import { meshGradientMeta, ShaderFit, ShaderFitOptions } from '@paper-design/shaders';
-import { useColors } from '@/helpers/use-colors';
+import { Controllers, ControlSchema, createFolder, renderControlSchema } from '@/components/controllers';
+import { ColorPicker } from '@/components/color-picker';
+import { useColors } from '@/helpers/use-colors-standalone';
 
 /**
  * You can copy/paste this example to use MeshGradient in your app
@@ -29,71 +28,133 @@ const MeshGradientWithControls = () => {
     maxColorCount: meshGradientMeta.maxColorCount,
   });
 
-  const [params, setParams] = useControls(() => {
-    return {
-      Parameters: folder(
-        {
-          distortion: { value: defaults.distortion, min: 0, max: 1, order: 200 },
-          swirl: { value: defaults.swirl, min: 0, max: 1, order: 201 },
-          speed: { value: defaults.speed, min: 0, max: 2, order: 400 },
-        },
-        { order: 1 }
-      ),
-      Transform: folder(
-        {
-          scale: { value: defaults.scale, min: 0.01, max: 4, order: 400 },
-          rotation: { value: defaults.rotation, min: 0, max: 360, order: 401 },
-          offsetX: { value: defaults.offsetX, min: -1, max: 1, order: 402 },
-          offsetY: { value: defaults.offsetY, min: -1, max: 1, order: 403 },
-        },
-        {
-          order: 2,
-          collapsed: false,
-        }
-      ),
-      Fit: folder(
-        {
-          fit: { value: defaults.fit, options: Object.keys(ShaderFitOptions) as ShaderFit[], order: 404 },
-          worldWidth: { value: 1000, min: 0, max: 5120, order: 405 },
-          worldHeight: { value: 500, min: 0, max: 5120, order: 406 },
-          originX: { value: defaults.originX, min: 0, max: 1, order: 407 },
-          originY: { value: defaults.originY, min: 0, max: 1, order: 408 },
-        },
-        {
-          order: 3,
-          collapsed: true,
-        }
-      ),
-    };
-  }, [colors.length]);
-
-  useControls(() => {
-    const presets = Object.fromEntries(
-      meshGradientPresets.map(({ name, params: { worldWidth, worldHeight, ...preset } }) => [
-        name,
-        button(() => {
-          const { colors, ...presetParams } = preset;
-          setColors(colors);
-          setParamsSafe(params, setParams, presetParams);
-        }),
-      ])
-    );
-    return {
-      Presets: folder(presets, { order: 10 }),
-    };
+  const [params, setParams] = useState({
+    distortion: defaults.distortion,
+    swirl: defaults.swirl,
+    speed: defaults.speed,
+    scale: defaults.scale,
+    rotation: defaults.rotation,
+    offsetX: defaults.offsetX,
+    offsetY: defaults.offsetY,
+    fit: defaults.fit,
+    worldWidth: worldWidth || 1000,
+    worldHeight: worldHeight || 500,
+    originX: defaults.originX,
+    originY: defaults.originY,
   });
 
-  // Reset to defaults on mount, so that Leva doesn't show values from other
-  // shaders when navigating (if two shaders have a color1 param for example)
-  useResetLevaParams(params, setParams, defaults);
-  usePresetHighlight(meshGradientPresets, params);
-  cleanUpLevaParams(params);
+  const updateParams = (updates: Partial<typeof params>) => {
+    setParams((prev) => ({ ...prev, ...updates }));
+  };
+
+  const controlSchema: ControlSchema = {
+    Colors: {
+      _isFolder: true,
+      config: {
+        colorCount: {
+          value: colors.length,
+          min: 1,
+          max: meshGradientMeta.maxColorCount,
+          step: 1,
+          onChange: (value) => {
+            const newColors = [...colors];
+            if (value > colors.length) {
+              for (let i = colors.length; i < value; i++) {
+                newColors.push(`hsla(${(40 * i) % 360}, 100%, 50%, 1)`);
+              }
+            } else {
+              newColors.length = value;
+            }
+            setColors(newColors);
+          },
+        },
+        colorGrid: {
+          type: 'custom',
+          render: () => (
+            <div className="grid grid-cols-2">
+              {colors.map((color, index) => (
+                <ColorPicker
+                  key={index}
+                  label=""
+                  value={color}
+                  onChange={(newColor) => {
+                    const newColors = [...colors];
+                    newColors[index] = newColor;
+                    setColors(newColors);
+                  }}
+                />
+              ))}
+            </div>
+          ),
+        },
+      },
+      options: {},
+    },
+    Parameters: createFolder(
+      {
+        distortion: { min: 0, max: 1 },
+        swirl: { min: 0, max: 1 },
+        speed: { min: 0, max: 2 },
+      },
+      params,
+      updateParams
+    ),
+    Transform: createFolder(
+      {
+        scale: { min: 0.01, max: 4 },
+        rotation: { min: 0, max: 360, step: 1 },
+        offsetX: { min: -1, max: 1 },
+        offsetY: { min: -1, max: 1 },
+      },
+      params,
+      updateParams,
+      { collapsed: false }
+    ),
+    Fit: createFolder(
+      {
+        fit: { options: Object.keys(ShaderFitOptions) as ShaderFit[] },
+        worldWidth: { min: 0, max: 5120, step: 1 },
+        worldHeight: { min: 0, max: 5120, step: 1 },
+        originX: { min: 0, max: 1 },
+        originY: { min: 0, max: 1 },
+      },
+      params,
+      updateParams,
+      { collapsed: true }
+    ),
+    Presets: createFolder(
+      Object.fromEntries(
+        meshGradientPresets.map(
+          ({ name, params: { worldWidth: presetWorldWidth, worldHeight: presetWorldHeight, ...preset } }) => [
+            name,
+            {
+              type: 'button' as const,
+              onClick: () => {
+                const { colors, ...presetParams } = preset;
+                setColors(colors);
+                updateParams({
+                  ...presetParams,
+                  worldWidth: presetWorldWidth || 1000,
+                  worldHeight: presetWorldHeight || 500,
+                });
+              },
+            },
+          ]
+        )
+      ),
+      params,
+      updateParams
+    ),
+  };
 
   return (
     <>
       <Link href="/">
         <BackButton />
       </Link>
+
+      <Controllers>{renderControlSchema(controlSchema)}</Controllers>
+
       <MeshGradient {...params} colors={colors} className="fixed size-full" />
     </>
   );
